fix(robustez): WFE consolidado (DQ Labs canônico) + casts np→float + grids finos
- wfa.py: WFE agora é CONSOLIDADO (ΣOOS/ΣIS×100), como o material DQ Labs
  cap. 07 define literalmente ("resultado consolidado de todos os períodos OOS
  pelo consolidado dos IS"). A versão anterior (média dos ratios por janela,
  zerando IS≤0 — herdada do WIP) era patológica: reprovava estratégia com OOS
  POSITIVO em todas as janelas. Diagnóstico no sweep v2: 6048 testados, 24
  passaram o screen de curva e TODOS morriam nesse gate — enquanto o gate
  honesto de "OOS negativas <50%" passava em 100% deles.
- backtest.py: float() em equity_bar e normalized_lot — np.float64 vazava pro
  JSON do publish/worker (json.dumps quebra com np types).
- miner.py: grids mais finos (sl 1.0-2.5 × tp 1.5-3.0; breakout lookback
  8-50) — mais espaço de busca legítimo pro funil, gates intactos.

+++ scripts/robustez/backtest.py
@@ -96,7 +96,8 @@ def normalized_lot(atr_mean: float, contract_size: float) -> float:
     if dollars_per_atr_per_lot <= 0:
         return 0.01
     lot = TARGET_ATR_RISK / dollars_per_atr_per_lot
-    return max(0.01, round(lot, 2))
+    # float() evita np.float64 vazar pro JSON (json.dumps quebra com np types)
+    return float(max(0.01, round(lot, 2)))
 
 
 def run_backtest(
@@ -188,7 +189,7 @@ def run_backtest(
 
         # equity mark-to-market da barra (realizado + posição aberta)
         floating = pos * (close[i] - entry_px) * value if pos != 0 else 0.0
-        equity_bar.append(round(realized + floating, 2))
+        equity_bar.append(float(round(realized + floating, 2)))
         equity_dates.append(str(pd.Timestamp(times[i]).date()))
 
     return BacktestResult(trades=trades, equity_bar=equity_bar,
@@ -262,7 +263,7 @@ def _run_grid(
                 basket, bdir = [px], -1
 
         floating = sum(bdir * (px - e) * value for e in basket) if basket else 0.0
-        equity_bar.append(round(realized + floating, 2))
+        equity_bar.append(float(round(realized + floating, 2)))
         equity_dates.append(str(pd.Timestamp(times[i]).date()))
 
     return BacktestResult(trades=trades, equity_bar=equity_bar,
